+++ tests/testthat/test-quasiquotation.R
@@ -149,6 +149,10 @@ test_that("LHS of nested `!!` is expanded (#405)", {
   expect_identical_(expr(!!1 + !!2 + foo(!!3) + !!4), quote(1 + 2 + foo(3) + 4))
 })
 
+test_that("operators with single argument work (#652)", {
+  expect_identical(expr(`/`(2)), quote(`/`(2)))
+})
+
 test_that("evaluates contents of `!!`", {
   expect_identical(expr(!!(1 + 2)), 3)
 })